Toggle separate circle color property options for when large map active or not.
Added the ability to have a toggle for separate color display properties for circles based on if the large overlay map is active or not and the new toggle setting is enabled. Example of use includes, setting different alpha values so that when the large overlay map is open, the game world circle is set to a lower alpha value to be less distracting. Refactored  old “Color” to “GameWorldColor” to be more descriptive now that there is more than 1 color property per item.